// src/common/enums/primeLocation.ts
export enum PrimeLocation {
  // Tallon Overworld
  LANDING_SITE = 'Landing Site',
  ALCOVE = 'Alcove',
  FRIGATE_CRASH_SITE = 'Frigate Crash Site',
  OVERGROWN_CAVERN = 'Overgrown Cavern',
  ROOT_CAVE = 'Root Cave',
  ARTIFACT_TEMPLE = 'Artifact Temple',
  TRANSPORT_TUNNEL_B = 'Transport Tunnel B',
  ARBOR_CHAMBER = 'Arbor Chamber',
  CARGO_FREIGHT_LIFT_TO_DECK_GAMMA = 'Cargo Freight Lift to Deck Gamma',
  BIOHAZARD_CONTAINMENT = 'Biohazard Containment',
  HYDRO_ACCESS_TUNNEL = 'Hydro Access Tunnel',
  GREAT_TREE_CHAMBER = 'Great Tree Chamber',
  LIFE_GROVE_TUNNEL = 'Life Grove Tunnel',
  LIFE_GROVE_START = 'Life Grove (Start)',
  LIFE_GROVE_UNDERWATER_SPINNER = 'Life Grove (Underwater Spinner)',

  // Chozo Ruins
  MAIN_PLAZA_HALF_PIPE = 'Main Plaza (Half-Pipe)',
  MAIN_PLAZA_GRAPPLE_LEDGE = 'Main Plaza (Grapple Ledge)',
  MAIN_PLAZA_TREE = 'Main Plaza (Tree)',
  MAIN_PLAZA_LOCKED_DOOR = 'Main Plaza (Locked Door)',
  RUINED_FOUNTAIN = 'Ruined Fountain',
  RUINED_SHRINE_BEETLE_BATTLE = 'Ruined Shrine (Beetle Battle)',
  RUINED_SHRINE_HALF_PIPE = 'Ruined Shrine (Half-Pipe)',
  RUINED_SHRINE_LOWER_TUNNEL = 'Ruined Shrine (Lower Tunnel)',
  VAULT = 'Vault',
  TRAINING_CHAMBER = 'Training Chamber',
  RUINED_NURSERY = 'Ruined Nursery',
  TRAINING_CHAMBER_ACCESS = 'Training Chamber Access',
  MAGMA_POOL = 'Magma Pool',
  TOWER_OF_LIGHT = 'Tower of Light',
  TOWER_CHAMBER = 'Tower Chamber',
  RUINED_GALLERY_MISSILE_WALL = 'Ruined Gallery (Missile Wall)',
  RUINED_GALLERY_TUNNEL = 'Ruined Gallery (Tunnel)',
  TRANSPORT_ACCESS_NORTH = 'Transport Access North',
  GATHERING_HALL = 'Gathering Hall',
  HIVE_TOTEM = 'Hive Totem',
  SUNCHAMBER_FLAAHGRA = 'Sunchamber (Flaahgra)',
  SUNCHAMBER_GHOSTS = 'Sunchamber (Ghosts)',
  WATERY_HALL_ACCESS = 'Watery Hall Access',
  WATERY_HALL_SCAN_PUZZLE = 'Watery Hall (Scan Puzzle)',
  WATERY_HALL_UNDERWATER = 'Watery Hall (Underwater)',
  DYNAMO_LOWER = 'Dynamo (Lower)',
  DYNAMO_SPIDER_TRACK = 'Dynamo (Spider Track)',
  BURN_DOME_TUNNEL = 'Burn Dome (Tunnel)',
  BURN_DOME_I_DRONE = 'Burn Dome (I. Drone)',
  FURNACE_SPIDER_TRACKS = 'Furnace (Spider Tracks)',
  FURNACE_TUNNEL = 'Furnace (Tunnel)',
  HALL_OF_THE_ELDERS = 'Hall of the Elders',
  CROSSWAY = 'Crossway',
  ELDER_CHAMBER = 'Elder Chamber',
  ANTECHAMBER = 'Antechamber',

  // Magmoor Caverns
  LAVA_LAKE = 'Lava Lake',
  TRICLOPS_PIT = 'Triclops Pit',
  STORAGE_CAVERN = 'Storage Cavern',
  TRANSPORT_TUNNEL_A = 'Transport Tunnel A',
  WARRIOR_SHRINE = 'Warrior Shrine',
  SHORE_TUNNEL = 'Shore Tunnel',
  FIERY_SHORES_MORPH_TRACK = 'Fiery Shores (Morph Track)',
  FIERY_SHORES_WARRIOR_SHRINE_TUNNEL = 'Fiery Shores (Warrior Shrine Tunnel)',
  PLASMA_PROCESSING = 'Plasma Processing',
  MAGMOOR_WORKSTATION = 'Magmoor Workstation',

  // Phendrana Drifts
  PHENDRANA_SHORELINES_BEHIND_ICE = 'Phendrana Shorelines (Behind Ice)',
  PHENDRANA_SHORELINES_SPIDER_TRACK = 'Phendrana Shorelines (Spider Track)',
  CHOZO_ICE_TEMPLE = 'Chozo Ice Temple',
  ICE_RUINS_WEST = 'Ice Ruins West',
  ICE_RUINS_EAST_BEHIND_ICE = 'Ice Ruins East (Behind Ice)',
  ICE_RUINS_EAST_SPIDER_TRACK = 'Ice Ruins East (Spider Track)',
  CHAPEL_OF_THE_ELDERS = 'Chapel of the Elders',
  RUINED_COURTYARD = 'Ruined Courtyard',
  PHENDRANA_CANYON = 'Phendrana Canyon',
  QUARANTINE_CAVE = 'Quarantine Cave',
  RESEARCH_LAB_HYDRA = 'Research Lab Hydra',
  QUARANTINE_MONITOR = 'Quarantine Monitor',
  OBSERVATORY = 'Observatory',
  TRANSPORT_ACCESS = 'Transport Access',
  CONTROL_TOWER = 'Control Tower',
  RESEARCH_CORE = 'Research Core',
  FROST_CAVE = 'Frost Cave',
  RESEARCH_LAB_AETHER_TANK = 'Research Lab Aether (Tank)',
  RESEARCH_LAB_AETHER_MORPH_TRACK = 'Research Lab Aether (Morph Track)',
  GRAVITY_CHAMBER_UNDERWATER = 'Gravity Chamber (Underwater)',
  GRAVITY_CHAMBER_GRAPPLE_LEDGE = 'Gravity Chamber (Grapple Ledge)',
  STORAGE_CAVE = 'Storage Cave',
  SECURITY_CAVE = 'Security Cave',

  // Phazon Mines
  MAIN_QUARRY = 'Main Quarry',
  SECURITY_ACCESS_A = 'Security Access A',
  STORAGE_DEPOT_B = 'Storage Depot B',
  STORAGE_DEPOT_A = 'Storage Depot A',
  ELITE_RESEARCH_PHAZON_ELITE = 'Elite Research (Phazon Elite)',
  ELITE_RESEARCH_LASER = 'Elite Research (Laser)',
  ELITE_CONTROL_ACCESS = 'Elite Control Access',
  VENTILATION_SHAFT = 'Ventilation Shaft',
  PHAZON_PROCESSING_CENTER = 'Phazon Processing Center',
  PROCESSING_CENTER_ACCESS = 'Processing Center Access',
  ELITE_QUARTERS = 'Elite Quarters',
  CENTRAL_DYNAMO = 'Central Dynamo',
  METROID_QUARANTINE_B = 'Metroid Quarantine B',
  METROID_QUARANTINE_A = 'Metroid Quarantine A',
  FUNGAL_HALL_B = 'Fungal Hall B',
  PHAZON_MINING_TUNNEL = 'Phazon Mining Tunnel',
  FUNGAL_HALL_ACCESS = 'Fungal Hall Access'
}